Patch video update in FCEUX_Update
In the original implementation of FCEUX_Update video and audio updates
are integrated, which could lead to a possible defect that audio
streaming blocks video streaming. It will try to fill up the sound
buffer before writing into the video buffer. Thus if the sound buffer is
consumed at a faster rate than writing, video is not going to be written
into, leading to a black screen.

Moving BlitScreen out of audio update could help ease this situation.
But this is a temporary patch which may still lead to further bugs since
video update is not separated from audio update.